backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
passlib[bcrypt]
motor
groq
sentence-transformers
pymupdf
python-multipart
